% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAccord_fns.R
\name{clone_df_to_cx_df}
\alias{clone_df_to_cx_df}
\title{Transform dataframes from mutations to exclusivity.}
\usage{
clone_df_to_cx_df(clone_df)
}
\arguments{
\item{clone_df}{The dataframe of mutational presence in each clone.
The first column must be the sample id, the next column an identifier
of each tree obtained per sample. The third column stores the frequency of each
clone. The remaining columns are for the genes/positions.
Each row is a clone (per sample and per tree) with its mutation status
over the genes recorded in binary with 1 representing mutation.}
}
\value{
A dataframe with one row per sample with columns for each pair
of genes recording the average (over trees) clonal exclusivity.
}
\description{
Turn a data frame of clonal mutations into a data frame of pairwise
clonal exclusivity patterns
}
\author{
Jack Kuipers
}
