% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAccord_fns.R
\name{filter_cx_df}
\alias{filter_cx_df}
\title{Filters a clonal exclusivity dataframe.}
\usage{
filter_cx_df(cx_df, k)
}
\arguments{
\item{cx_df}{A dataframe with one row per sample and columns for each
pair of genes recording the average (over trees) clonal exclusivity.}

\item{k}{The minimum of samples to contain}
}
\value{
A filtered dataframe of clonal exclusivity patterns with an additional
first column of clonal exclusivity probabilities per sample.
}
\description{
Filter a data frame of pairwise clonal exclusivity patterns
to select samples with a clonal exclusivity probability strictly
between 0 and 1 and gene pairs which occur in at least a given
number of samples.
}
\author{
Jack Kuipers
}
