% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAccord_fns.R
\name{geneAccord}
\alias{geneAccord}
\title{Perform the geneAccord clonal exclusivity test .}
\usage{
geneAccord(
  clonal_exclusivity_df,
  exact_limit = 12,
  filter = TRUE,
  MC_reps = NULL
)
}
\arguments{
\item{clonal_exclusivity_df}{A data frame with one row per sample with
the first column containing the and following columns for each pair
of genes recording the average (over trees) clonal exclusivity.}

\item{exact_limit}{Limit above which the asymptotic chi-squared distribution
is employed instead of the exact test which is more computationally intensive.}

\item{filter}{Whether to remove untestable gene pairs (TRUE, default) to
reduce multiple testing}

\item{MC_reps}{How many repetitions to use for Monte Carlo estimation of
the p-values. When NULL (default) exact or chi-squared testing is used.}
}
\value{
A data fram with the gene pairs, number of observed samples containing
the gene pair, the number of samples with clonal exclusivity, the maximum
likelihood estimate of delta, the LR test statistic of the clonal exclusivity
test and the corresponding p-value and BH corrected q-value.
}
\description{
Takes a data frame of clonal exclusivity patterns with rows representing
patient sample, a first column giving the overall observed exclusivity
probability and subsequent columns providing the rate of clonal
exclusivity of the tree(s) of that patient sample.
}
\author{
Jack Kuipers
}
